+++ src/extension.ts
@@ -2,11 +2,8 @@
 
 import * as vscode from 'vscode';
 import { project } from "./project";
-import { Utils } from "./Utils";
 
 export async function activate(context: vscode.ExtensionContext) {
-    await Utils.loadPackageInfo(context);
-
     let createCProjectCommand = vscode.commands.registerCommand('extension.createCProject', () => {
         project.createProject("c");
     });